separate git and config functionality
