+++ pkg/plugins/platform/utils.go
@@ -4,13 +4,10 @@ import (
 	"bytes"
 	"fmt"
 	"html/template"
-	"io/ioutil"
 	"os"
 	"path"
 	"strings"
 	"time"
-
-	"gopkg.in/yaml.v2"
 
 	log "github.com/Sirupsen/logrus"
 
@@ -62,49 +59,6 @@ func WriteLocalAppYAML(app App) error {
 		return err
 	}
 	f.WriteString(rendered)
-	return nil
-}
-
-// CloneSource clones or pulls a repo
-func CloneSource(app App) error {
-
-	details, err := app.GetRepoDetails()
-	if err != nil {
-		return err
-	}
-
-	cloneURL, err := details.GetCloneURL()
-	if err != nil {
-		return err
-	}
-	cfg, _ := GetConfig()
-	basePath := path.Join(cfg.Workspace, app.RelPath(), "src")
-
-	out, err := system.RunCommand("git", []string{
-		"clone", "-b", details.Branch, cloneURL, basePath,
-	})
-	if err != nil {
-		if !strings.Contains(string(out), "already exists") {
-			return fmt.Errorf("%s - %s", err.Error(), string(out))
-		}
-
-		fmt.Print("Local copy of site exists, updating... ")
-
-		out, err = system.RunCommand("git", []string{
-			"-C", basePath,
-			"pull", "origin", details.Branch,
-		})
-		if err != nil {
-			return fmt.Errorf("%s - %s", err.Error(), string(out))
-		}
-
-		fmt.Printf("Updated to latest in %s branch\n", details.Branch)
-	}
-
-	if len(out) > 0 {
-		log.Info(string(out))
-	}
-
 	return nil
 }
 
@@ -366,117 +320,9 @@ func SubTag(image string, tag string) string {
 	return strings.Join(parts, ":")
 }
 
-// Config represents the data that is or can be stored in $HOME/.drud
-type Config struct {
-	APIVersion      string `yaml:"apiversion"`
-	ActiveApp       string `yaml:"activeapp"`
-	ActiveDeploy    string `yaml:"activedeploy"`
-	Client          string `yaml:"client"`
-	DrudHost        string `yaml:"drudhost"`
-	GithubAuthToken string `yaml:"githubauthtoken"`
-	GithubAuthOrg   string `yaml:"githubauthorg"`
-	Protocol        string `yaml:"protocol"`
-	VaultAddr       string `yaml:"vaultaddr"`
-	VaultAuthToken  string `yaml:"vaultauthtoken"`
-	Workspace       string `yaml:"workspace"`
-}
-
 // EveHost creates the eve host string from the config.
 func (cfg *Config) EveHost() string {
 	return fmt.Sprintf("%s://%s/%s/", cfg.Protocol, cfg.DrudHost, cfg.APIVersion)
-}
-
-func parseConfigFlag() string {
-	var value string
-
-	for i, arg := range os.Args {
-		if strings.HasPrefix(arg, "--config=") {
-			value = strings.TrimPrefix(arg, "--config=")
-		} else if arg == "--config" {
-			value = os.Args[i+1]
-		}
-	}
-	if value == "" {
-		home, _ := system.GetHomeDir()
-		value = fmt.Sprintf("%v/drud.yaml", home)
-	}
-
-	if _, err := os.Stat(value); os.IsNotExist(err) {
-		var cFile, err = os.Create(value)
-		if err != nil {
-			log.Fatal(err)
-		}
-		cFile.Close()
-	}
-	return value
-}
-
-// GetConfig Loads a config structure from yaml and environment.
-func GetConfig() (cfg *Config, err error) {
-	cfgFile := parseConfigFlag()
-
-	source, err := ioutil.ReadFile(cfgFile)
-	if err != nil {
-		panic(err)
-	}
-
-	c := &Config{}
-	err = yaml.Unmarshal(source, c)
-	if err != nil {
-		panic(err)
-	}
-
-	if c.APIVersion == "" && os.Getenv("DRUD_APIVERSION") != "" {
-		c.APIVersion = os.Getenv("DRUD_APIVERSION")
-	}
-	if c.ActiveApp == "" && os.Getenv("DRUD_ACTIVEAPP") != "" {
-		c.ActiveApp = os.Getenv("DRUD_ACTIVEAPP")
-	}
-	if c.ActiveDeploy == "" && os.Getenv("DRUD_ACTIVEDEPLOY") != "" {
-		c.ActiveDeploy = os.Getenv("DRUD_ACTIVEDEPLOY")
-	}
-	if c.Client == "" && os.Getenv("DRUD_CLIENT") != "" {
-		c.Client = os.Getenv("DRUD_CLIENT")
-	}
-	if c.DrudHost == "" && os.Getenv("DRUD_DRUDHOST") != "" {
-		c.DrudHost = os.Getenv("DRUD_DRUDHOST")
-	}
-	if c.GithubAuthToken == "" && os.Getenv("DRUD_GITHUBAUTHTOKEN") != "" {
-		c.GithubAuthToken = os.Getenv("DRUD_GITHUBAUTHTOKEN")
-	}
-	if c.GithubAuthOrg == "" && os.Getenv("DRUD_GITHUBAUTHORG") != "" {
-		c.GithubAuthOrg = os.Getenv("DRUD_GITHUBAUTHORG")
-	}
-	if c.Protocol == "" && os.Getenv("DRUD_PROTOCOL") != "" {
-		c.Protocol = os.Getenv("DRUD_PROTOCOL")
-	}
-	if c.VaultAddr == "" && os.Getenv("DRUD_VAULTADDR") != "" {
-		c.VaultAddr = os.Getenv("DRUD_VAULTADDR")
-	}
-	if c.VaultAuthToken == "" && os.Getenv("DRUD_VAULTAUTHTOKEN") != "" {
-		c.VaultAuthToken = os.Getenv("DRUD_VAULTAUTHTOKEN")
-	}
-	if c.Workspace == "" && os.Getenv("DRUD_WORKSPACE") != "" {
-		c.Workspace = os.Getenv("DRUD_WORKSPACE")
-	}
-
-	return c, nil
-}
-
-// WriteConfig writes each config value to the BoltDB and updates the
-// global cfg as well.
-func (cfg *Config) WriteConfig(f string) (err error) {
-	cfgbytes, err := yaml.Marshal(cfg)
-	if err != nil {
-		return err
-	}
-
-	err = ioutil.WriteFile(f, cfgbytes, 0644)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
 
 func ComposeFileExists(app App) bool {
